Decrease size of used integers, get rid of SafeMath

+++ contracts/StakingAccessInterface.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.6.0;
 contract StakingAccessInterface {
-    function moveToStake(address user, uint256 amount) public {
+    function moveToStake(address user, uint64 amount) public {
     }
-    function moveFromStake(address user, uint256 amount) public {
+    function moveFromStake(address user, uint64 amount) public {
     }
 }
